Split implementation to multiple files.
The dsl and functions implementations are conceptually independent, so split
them out.
The main hstore implementation is also separated.
